pyproject.toml
[tool.poetry]
name = "friskis"
version = "0.1.0"
description = ""
authors = ["David <davidsvenson@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.1"
requests = "^2.23.0"
pytz = "^2019.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
